requirements.txt
PyPDF2==3.0.1
pdf2image==1.16.3
python-docx==1.0.0
pytesseract==0.3.10
python-magic==0.4.27